pyproject.toml
[project]
name = "seer"
version = "0.3.0"
description = "Modular infrastructure for interpretability experiments with agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "ajobi-uhc"}
]
keywords = ["interpretability", "AI", "agents", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0",
    "claude-agent-sdk>=0.1.0",
    "fastmcp>=2.10.6",
    "ipykernel>=6.30.0",
    "jupyter-client>=8.6.3",
    "jupyter-server>=2.16.0",
    "nbformat>=5.10.4",
    "modal>=0.49.0",
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "openai-agents>=0.4.0",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0",
    "requests>=2.32.4",
    "rich>=13.0.0",
    "skypilot-nightly[runpod]",
    "runpod>=1.5.1",
    "matplotlib>=3.10.7"
]

[project.urls]
Homepage = "https://github.com/ajobi-uhc/seer"
Documentation = "https://ajobi-uhc.github.io/seer/"
Repository = "https://github.com/ajobi-uhc/seer"
Issues = "https://github.com/ajobi-uhc/seer/issues"

[project.optional-dependencies]
web = [
    "flask>=3.0.0"
]
docs = [
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0"
]

[project.scripts]
scribe = "scribe.cli.cli:main"
modal-sandbox-mcp = "modal_sandbox_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["scribe*", "src*", "modal_sandbox_mcp*"]
